We really don't want a file called diffTerm.pdf in the repo.
Reason: if diffTerm.pdf is in the repo, everytime we recompile the
document on our local machine, and then try to pull changes from
the remote repo, git will complain that the remote (presumably old)
version of diffTerm.pdf will clobber our newly compiled version of
diffTerm.pdf.  Therefore, if we want to keep a compiled pdf version
of the paper in the repo, we should name it something like

    diffTerm-draft-20180501.pdf

Similarly, inputs/refs.bib should not be in the repo since it's
meant to be autogenerated when we compile the main tex document.