Please don't add the input/refs.bib file to the repo.
REASON: it should be automatically generated when we compile diffTerms.tex.

EXPLANATION: Here's how we deal with bibliography entries.
If you want to add a new bibtex entry, please add it to the
top of the main diffTerm.tex file.  Then REMOVE THE FILE
`inputs/refs.bib` (if it exists) so that next time you compile
diffTerm.tex (e.g., with pdflatex diffTerm.tex) this will
automatically generate a new input/refs.bib file that includes
all the bibtex entries at the top of diffTerm.tex.