(Sec 5, 6) Attempting to fix major flaws
I realized we don't want to generate the (possibly very large)
sets T \eta and T \zeta. I thought I saw a way around this, but
now I see another problem.  I'm in the midst of working on a solution
but making this commit because the algorithms are cleaner now...
though the complexity analysis section still needs to be edited
to account for the algorithm modifications.